% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMat.R
\name{calculateDistMatOneModel}
\alias{calculateDistMatOneModel}
\title{Calculate distance matrix with in one model}
\usage{
calculateDistMatOneModel(
  matrix,
  metadata,
  verbose = TRUE,
  method = "voom",
  subset.row = NULL,
  additional.variate = "donor"
)
}
\arguments{
\item{matrix}{matrix}

\item{metadata}{metadata}

\item{verbose}{print the message and progress bar (default: TRUE)}

\item{method}{voom or trend}

\item{subset.row}{subset.row}

\item{additional.variate}{additional.variate}
}
\value{
a list
}
\description{
description
}
\author{
Zhiyuan Hu
}
