% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Calculate cosine simialrity}
\usage{
cosine_similarity(Xr)
}
\arguments{
\item{Xr}{a matrix}
}
\description{
Refer to https://gist.github.com/bobthecat/2903031
}
