% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosineSimilarityR}
\alias{cosineSimilarityR}
\title{cosine similarity in R}
\usage{
cosineSimilarityR(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a similarity matrix among all rows of the input matrix
}
\description{
cosine similarity in R
}
