% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendritic}
\alias{dendritic}
\title{Seurat object of dendritic cells}
\format{
A data frame with xxx rows and xxx variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84133}
}
\usage{
dendritic
}
\description{
A Seurat object of dendritic cells
}
\examples{
data("dendritic")
}
\keyword{datasets}
