% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsampling.R
\name{downsampling}
\alias{downsampling}
\title{Downsampling}
\usage{
downsampling(
  metadata,
  n.size = 50,
  seed = 12345,
  include = FALSE,
  replace = FALSE,
  lower.cutoff = 3
)
}
\arguments{
\item{metadata}{metadata}

\item{n.size}{n.size (default: 50)}

\item{seed}{seed (default: 12345)}

\item{include}{using include = TRUE if including the cluster smaller than required size}

\item{replace}{using replace = TRUE if the cluster is smaller than required size}

\item{lower.cutoff}{lower cutoff}
}
\value{
return obj
}
\description{
description
}
\author{
Zhiyuan Hu
}
