% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{estimateProb}
\alias{estimateProb}
\title{Estimate probability}
\usage{
estimateProb(
  seu,
  ider,
  n_size = 40,
  seeds = c(12345, 89465, 10385, 10385, 17396),
  n.perm = 5,
  verbose = FALSE
)
}
\arguments{
\item{seu}{seu}

\item{ider}{ider}

\item{n_size}{40}
}
\description{
Estimate probability
}
