% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalClustering.R
\name{finalClustering}
\alias{finalClustering}
\title{Final clustering}
\usage{
finalClustering(seu, dist, cutree_by = "h", cutree.h = 0.35, cutree.k = 3)
}
\arguments{
\item{seu}{Seurat S4 object}

\item{dist}{dist matrix}
}
\value{
a list
}
\description{
description
}
\author{
Zhiyuan Hu
}
