% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeInitialClusters.R
\name{gatherInitialClusters}
\alias{gatherInitialClusters}
\title{Gather initial cluster names}
\usage{
gatherInitialClusters(seu_list, seu)
}
\arguments{
\item{seu_list}{seu_lisst}

\item{seu}{seu}
}
\value{
a seurat object
}
\description{
Gather initial cluster names
}
