% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDEr.R
\name{getGroupFit}
\alias{getGroupFit}
\title{Calculate Fit}
\usage{
getGroupFit(dge, design, contrast_m, method)
}
\arguments{
\item{dge}{dge}

\item{design}{design}

\item{contrast_m}{contrast_m}

\item{method}{method}
}
\description{
Calculate Fit
}
\seealso{
\code{\link{getIDEr}}
}
