% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDEr.R
\name{getIDEr}
\alias{getIDEr}
\title{Get IDEr matrix}
\usage{
getIDEr(seu, verbose = TRUE, downsampling.factor = 1, methods = "voom")
}
\arguments{
\item{seu}{Seurat S4 object}

\item{verbose}{print the message and progress bar (default: TRUE)}

\item{method}{voom or trend}
}
\value{
a list
}
\description{
description
}
\author{
Zhiyuan Hu
}
