% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{hdbscan.seurat}
\alias{hdbscan.seurat}
\title{Compute IDER-based similarity}
\usage{
hdbscan.seurat(seu, reduction = "pca", dims = 1:15, minPts = 25)
}
\arguments{
\item{seu}{seu after integration}

\item{reduction}{reduction}

\item{dims}{dims}

\item{minPts}{minPts}
}
\value{
A list of four values: three similarity matrices and one list of indeces.
}
\description{
Calculate the similarity matrix based on the metrics of Inter-group Differential ExpRession (IDER)
}
\details{
Details
}
\seealso{
\code{\link{plotNetwork}} \code{\link{finalClustering}}
}
