% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialClustering.R
\name{initialClustering}
\alias{initialClustering}
\title{Initial clustering}
\usage{
initialClustering(
  seu,
  batch.var = "Batch",
  cut.height = 0.4,
  nfeatures = 2000,
  additional.vars.to.regress = NULL,
  dims = 1:14,
  resolution = 0.6,
  downsampling.size = 50,
  verbose = FALSE
)
}
\arguments{
\item{seu}{Seurat S4 object. Required.}

\item{batch.var}{batch.var}

\item{cut.height}{cut.height}

\item{nfeatures}{nfeatures = 2000}

\item{dims}{dims}

\item{resolution}{resolution}

\item{downsampling.size}{downsampling.size}

\item{verbose}{verbose}

\item{additional.vars.to.regres}{additional.vars.to.regres}
}
\value{
Seurat S4 object with initial clustering results in `initial_cluster` of meta.data.
}
\description{
Perform batch-specific initial clustering.
}
