% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{measureSimilarity}
\alias{measureSimilarity}
\title{Measure similarity between two vectors}
\usage{
measureSimilarity(x1, x2, method = "pearson")
}
\arguments{
\item{x1}{x1}

\item{x2}{x2}

\item{method}{method}
}
\value{
similarity matrix
}
\description{
Measure similarity between two vectors
}
