% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeInitialClusters.R
\name{gatherInitialClusters}
\alias{gatherInitialClusters}
\title{Gather initial cluster names}
\usage{
gatherInitialClusters(seu_list, seu)
}
\arguments{
\item{seu_list}{A list containing Seurat objects. Required.}

\item{seu}{A Seurat object}
}
\value{
A Seurat object containing initial clustering  results in
`seu$initial_cluster`.
}
\description{
Gather initial cluster names
}
\section{Functions}{
\itemize{
\item \code{gatherInitialClusters}: initial clustering results from a Seurat object list to one
Seurat object. Follows the function `mergeInitialClusters`.
}}

\seealso{
\code{\link{mergeInitialClusters}}
}
