% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeInitialClusters.R
\name{mergeInitialClusters}
\alias{mergeInitialClusters}
\title{Merge Initial Clusters}
\usage{
mergeInitialClusters(
  seu_list,
  dist_list,
  use = "coef",
  method = "hc",
  hc.method = "average",
  cutree.by = "h",
  cutree.h = 0.6,
  cutree.k = 3
)
}
\arguments{
\item{seu_list}{seu_list}

\item{dist_list}{dist_list}

\item{use}{use}

\item{method}{method}

\item{hc.method}{hc.method}

\item{cutree.by}{cutree.by}

\item{cutree.h}{cutree.h}

\item{cutree.k}{cutree.k}
}
\value{
a list of seurat objects
}
\description{
Merge Initial Clusters
}
