% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreas}
\alias{pancreas}
\title{Pancreatic scRNA-Seq data.}
\format{
A Seurat object.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84133}
}
\usage{
pancreas
}
\description{
Toy data to test functions. It contains 12474 genes and only 222 cells.
The count matrix and sample information were downloaded from NCBI GEO
accession GSE84133.
}
\examples{
data("pancreas")
}
\keyword{datasets}
