% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreas_counts}
\alias{pancreas_counts}
\title{Sparse count matrix of pancreatic cells}
\format{
A data frame with xxx rows and xxx variables:
\describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84133}
}
\usage{
pancreas_counts
}
\description{
A count matrix containing xxx
}
\keyword{datasets}
