% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreas_meta}
\alias{pancreas_meta}
\title{Metadata of pancreatic cells}
\format{
A data frame with xxx rows and xxx variables:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE84133}
}
\usage{
pancreas_meta
}
\description{
A data.frame containing cell annotations
}
\keyword{datasets}
