% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistMat.R
\name{plotDistMat}
\alias{plotDistMat}
\title{Plot Similarity Matrix}
\usage{
plotDistMat(dist.list, use = "coef")
}
\arguments{
\item{dist.list}{List of list of similarity matrix. Output of 
function `getDistMat()`. Required.}

\item{use}{Choose from "coef" and "p". (Default: coef)}
}
\value{
a pheatmap showing the similarity matrix
}
\description{
description
}
\seealso{
\code{\link{getDistMat}}
}
