% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot Network Graph}
\usage{
plotNetwork(
  seu,
  dist,
  col.vector = NULL,
  use = "coef",
  colour.by = NULL,
  weight.factor = 6.5,
  vertex.size = 1
)
}
\arguments{
\item{seu}{Seurat S4 object after the step of `getIDEr`. Required.}

\item{dist}{A list. Output of `getIDEr`. Required.}

\item{col.vector}{A vector of Hex colour codes. If no value is given, a vector of 74 colours will be used. (Default: NULL)}

\item{use}{Character. The type of dist matrix to use. Can be one of "coef", "t" and "p". (Default: coef)}

\item{colour.by}{Character. It should be one of the colnames of Seurat object meta.data. It is used to colour the vertex of the networkgraph. (Default: NULL)}

\item{weight.factor}{Numerical. Adjust the thickness of the edges. (Default: 6.5)}

\item{vertex.size}{Numerical. Adjsut the size of vertexes. (Default: 1)}
}
\value{
An igraph object
}
\description{
Plot network for the initial clusters based on IDEr. The width of edges denotes the similarity between two initial clusters.
}
\details{
Details
}
\seealso{
\code{\link{getIDEr}} \code{\link[igraph]{graph_from_data_frame}}
}
