% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Plot Network Graph}
\usage{
scatterPlot(
  seu,
  reduction,
  colour.by,
  colvec = NULL,
  title = NULL,
  sort.by.numbers = TRUE,
  viridis_option = "B"
)
}
\arguments{
\item{seu}{Seurat S4 object after the step of `getIDER`. Required.}

\item{reduction}{list. Output of `getIDER`. Required.}

\item{colour.by}{A vector of Hex colour codes. If no value is given, a vector of 74 colours will be used. (Default: NULL)}

\item{colvec}{Character. The type of dist matrix to use. Can be one of "coef", "t" and "p". (Default: coef)}

\item{title}{Character. It should be one of the colnames of Seurat object meta.data. It is used to colour the vertex of the networkgraph. (Default: NULL)}

\item{sort.by.numbers}{Numerical. Adjust the thickness of the edges. (Default: 6.5)}
}
\value{
An igraph object
}
\description{
Plot network for the initial clusters based on IDEr. The width of edges denotes the similarity between two initial clusters.
}
\details{
Details
}
