% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMat.R
\name{calculateDistMatOneModel}
\alias{calculateDistMatOneModel}
\title{Calculate distance matrix with in one model}
\usage{
calculateDistMatOneModel(
  matrix,
  metadata,
  verbose = TRUE,
  method = "voom",
  additional.variate = NULL
)
}
\arguments{
\item{matrix}{The count matrix. Rows are genes/features and columns are
samples/cells.}

\item{metadata}{Data frame. Its rows should correspond to
columns of the `matrix` input.}

\item{verbose}{Print the message and progress bar (default: TRUE)}

\item{method}{Methods for DE analysis. Options: "voom" or "trend" (default)}

\item{additional.variate}{additional variate to include into the linear
model to regress out}
}
\value{
A similarity matrix
}
\description{
This function is called by `getDistMat`.
}
\seealso{
This function is called by \code{\link{getDistMat}}
}
\author{
Zhiyuan Hu
}
