% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{dendritic}
\alias{dendritic}
\title{Dendritic Data}
\format{
A Seurat object.
}
\source{
The data were downloaded from \url{https://hub.docker.com/r/jinmiaochenlab/batch-effect-removal-benchmarking}.  
Reference 1: Tran HTN, Ang KS, Chevrier M, Lee NYS, Goh M, Chen J. A benchmark of batch-effect correction methods for single-cell RNA sequencing data. Genome Biol. 2020;21:1–32.  
Reference 2: Villani A-C, Satija R, Reynolds G, Sarkizova S, Shekhar K, Fletcher J, et al. Single-cell RNA-seq reveals new types of human blood dendritic cells, monocytes, and progenitors. Science. 2017;356(6335).
}
\usage{
data(dendritic)
}
\description{
This dataset comprises data for four dendritic cell subtypes (CD141, 
CD1C, DoubleNeg, and pDC) from two batches. The raw count matrix and sample 
information were downloaded from a curated set, and cells with fewer than 
500 detected genes have been removed.
}
\details{
A Seurat object containing dendritic cell data.
}
\keyword{datasets}
