% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{dendritic}
\alias{dendritic}
\title{Dendritic Data}
\format{
A Seurat Object
}
\source{
The data were downloaded from https://hub.docker.com/r/jinmiaochenlab/batch-effect-removal-benchmarking
ref 1: Tran HTN, Ang KS, Chevrier M, Lee NYS, Goh M, Chen J. A benchmark of batch-effect correction methods for single-cell RNA sequencing data. Genome Biol. 2020;21:1–32.
ref 2: Villani A-C, Satija R, Reynolds G, Sarkizova S, Shekhar K, Fletcher J, et al. Single-cell RNA-seq reveals new types of human blood dendritic cells, monocytes, and progenitors. Science. 2017;356(6335).
}
\usage{
data(dendritic)
}
\description{
The data contained four cell subtypes (CD141, CD1C, DoubleNeg, and pDC) from two batches.
The raw count matrix and the sample information were also downloaded from the curated set.
Cells with less than 500 genes detected were removed.
}
\keyword{datasets}
