% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeInitialClusters.R
\name{gatherInitialClusters}
\alias{gatherInitialClusters}
\title{Gather Initial Cluster Names}
\usage{
gatherInitialClusters(seu_list, seu)
}
\arguments{
\item{seu_list}{A list containing Seurat objects with initial clustering results. Required.}

\item{seu}{A Seurat object to which the merged initial cluster information will be added.}
}
\value{
A Seurat object containing the initial clustering results in the \code{initial_cluster} column of its \code{meta.data}.
}
\description{
Merge initial clustering results from a list of Seurat objects into a single Seurat object.
}
\seealso{
\code{\link{mergeInitialClusters}}
}
