% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMat.R
\name{getDistMat}
\alias{getDistMat}
\title{Calculate Distance Matrix}
\usage{
getDistMat(matrix, metadata, verbose = TRUE, method = "voom", model = "one")
}
\arguments{
\item{matrix}{matrix}

\item{metadata}{metadata}

\item{verbose}{print the message and progress bar (default: TRUE)}

\item{method}{voom or trend}

\item{model}{one model}
}
\value{
A list containing
}
\description{
discription
}
\author{
Zhiyuan Hu
}
