% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDistMat.R
\name{getDistMat}
\alias{getDistMat}
\title{Calculate Distance Matrix}
\usage{
getDistMat(
  seu_list,
  verbose = TRUE,
  tmp.initial.clusters = "seurat_clusters",
  method = "trend",
  additional.variate = NULL,
  downsampling.size = 35,
  downsampling.include = TRUE,
  downsampling.replace = TRUE
)
}
\arguments{
\item{seu_list}{seu list}

\item{verbose}{print the message and progress bar (default: TRUE)}

\item{method}{voom or trend}

\item{additional.variate}{additional variate to include into the regression model}

\item{downsampling.size}{downsampling.size}

\item{downsampling.include}{downsampling.include}

\item{downsampling.replace}{downsampling.replace}
}
\value{
a distance matrix
}
\description{
discription
}
\author{
Zhiyuan Hu
}
