% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDER.R
\name{getGroupFit}
\alias{getGroupFit}
\title{Calculate IDER-Based Similarity Between Two Groups}
\usage{
getGroupFit(logCPM, design, contrast_m)
}
\arguments{
\item{logCPM}{A numeric matrix of log-transformed counts per million.}

\item{design}{A design matrix for the differential expression analysis.}

\item{contrast_m}{A contrast matrix specifying the comparison between the two groups.}
}
\value{
A numeric value representing the IDER-based similarity between the two groups.
}
\description{
This function calculates the IDER-based similarity between two groups using a linear model.
}
