% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDER.R
\name{getGroupFit}
\alias{getGroupFit}
\title{Calculate linear regression Fit}
\usage{
getGroupFit(dge, design, contrast_m, method)
}
\arguments{
\item{dge}{dge}

\item{design}{design}

\item{contrast_m}{contrast matrix}

\item{method}{DE.method}
}
\description{
Calculate linear regression Fit
}
\seealso{
\code{\link{getIDEr}}
}
