% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDER.R
\name{getGroupFit}
\alias{getGroupFit}
\title{Calculate linear regression Fit}
\usage{
getGroupFit(logCPM, design, contrast_m)
}
\arguments{
\item{logCPM}{logCPM}

\item{design}{design}

\item{contrast_m}{contrast matrix}
}
\description{
Calculate linear regression Fit
}
