% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIDER.R
\name{getGroupFit}
\alias{getGroupFit}
\title{Calculate IDER-based similarity between two groups}
\usage{
getGroupFit(logCPM, design, contrast_m)
}
\arguments{
\item{logCPM}{logCPM}

\item{design}{design}

\item{contrast_m}{contrast matrix}
}
\value{
Numeric. The IDER-based similarity between two groups.
}
\description{
Calculate IDER-based similarity between two groups
}
