% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{pancreas_counts}
\alias{pancreas_counts}
\title{Pancreas Count Matrix}
\format{
A sparse matrix of class \code{"dgCMatrix"} with 12474 rows (genes) and 10127 columns (cells).
}
\source{
The count matrix and sample information were downloaded from NCBI GEO accession GSE84133.  
Reference: Baron M, Veres A, Wolock SL, Faust AL, Gaujoux R, Vetere 
A, et al. A single-cell transcriptomic map of the human and mouse pancreas reveals inter- 
and intra-cell population structure. Cell Syst. 2016;3:346–360.e4.
}
\usage{
data(pancreas_counts)
}
\description{
This dataset comprises a single-cell RNA-seq count matrix for pancreatic data from human and mouse samples.
}
\details{
A sparse count matrix for cross-species pancreatic data.
}
\keyword{datasets}
