% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{pancreas_meta}
\alias{pancreas_meta}
\title{Pancreas Metadata}
\format{
A data frame with 10127 rows and 3 columns:
\describe{
  \item{Batch}{Species information (human or mouse).}
  \item{Group}{Cell type annotation.}
  \item{Sample}{Donor information.}
}
}
\source{
The metadata were downloaded alongside the count matrix from NCBI GEO accession GSE84133.  
Reference: Baron M, Veres A, Wolock SL, Faust AL, Gaujoux R, Vetere A, et al. 
A single-cell transcriptomic map of the human and mouse pancreas reveals inter- and intra-cell 
population structure. Cell Syst. 2016;3:346–360.e4.
}
\usage{
data(pancreas_meta)
}
\description{
This dataset provides cell-level metadata for the human and mouse 
pancreatic data used in the study.
}
\details{
Cell-level metadata for cross-species pancreatic data.
}
\keyword{datasets}
