% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{pancreas_meta}
\alias{pancreas_meta}
\title{Pancreas Metadata}
\format{
A data frame with 10127 rows and 3 columns:
\describe{
  \item{Batch}{Species information, i.e., human or mouse}
  \item{Group}{Cell type annotation}
  \item{Sample}{Donor information}
}
}
\source{
The count matrix and sample information were downloaded from NCBI GEO accession GSE84133. 
ref: Baron M, Veres A, Wolock SL, Faust AL, Gaujoux R, Vetere A, et al. A single-cell transcriptomic map of the human and mouse pancreas reveals inter- and intra-cell population structure. Cell Syst. 2016;3:346–360.e4.
}
\usage{
data(pancreas_meta)
}
\description{
Cell-level metadata for human and mouse cross-species pancreatic data
}
\keyword{datasets}
