% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistMat.R
\name{plotDistMat}
\alias{plotDistMat}
\title{Plot Similarity Matrix with pheatmap}
\usage{
plotDistMat(dist.list, use = "coef")
}
\arguments{
\item{dist.list}{A list representing the similarity matrix output by \code{getDistMat()}. Required.}

\item{use}{Character string specifying the similarity measure to use. Default is "coef". 
No other option is currently available.}
}
\value{
A \code{pheatmap} object displaying the similarity matrix.
}
\description{
This function creates a heatmap of the similarity matrix computed by \code{getDistMat()}.
}
\seealso{
\code{\link{getDistMat}}
}
