% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistMat.R
\name{plotDistMat}
\alias{plotDistMat}
\title{Plot Similarity Matrix with pheatmap}
\usage{
plotDistMat(dist.list, use = "coef")
}
\arguments{
\item{dist.list}{Output of function `getDistMat()`. Required.}

\item{use}{Default: "coef". No other option currently that can be used.}
}
\value{
A pheatmap showing the similarity matrix
}
\description{
Plot Similarity Matrix with pheatmap
}
\seealso{
\code{\link{getDistMat}}
}
