% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistMat.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap for the IDER-based similarity matrix}
\usage{
plotHeatmap(seu, ider)
}
\arguments{
\item{seu}{Seurat object.}

\item{ider}{output of function `getIDEr`.}
}
\description{
description
}
\examples{
\donttest{
  plotHeatmap(seu, ider)
} 
}
\seealso{
\code{\link{getIDEr}}
}
