% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistMat.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap for the IDER-based similarity matrix}
\usage{
plotHeatmap(seu, ider, batch.var = "Batch")
}
\arguments{
\item{seu}{An Seurat object.}

\item{ider}{Output of function `getIDEr`.}

\item{batch.var}{Character. Metadata colname containing batch information.
(Default: Batch)}
}
\value{
A heatmap shows the similarity between shared groups in two batches
}
\description{
Plot Heatmap for the IDER-based similarity matrix
}
\examples{
\dontrun{
  plotHeatmap(seu, ider)
}
}
\seealso{
\code{\link{getIDEr}}
}
